requirements.txt
fastapi
uvicorn
pandas
Jinja2==3.0.3
fastapi-sqlalchemy
requests
psycopg2
python-multipart
tensorflow
Pillow
keras
h5py

aiofiles
asyncio